third_party/blink/web_tests/external/wpt/css/css-text/white-space/seg-break-transformation-002.tentative-expected.txt

This is a testharness.js-based test.
FAIL linebreak only assert_equals: expected 216 but got 223
FAIL spaces linebreak assert_equals: expected 216 but got 223
FAIL linebreak spaces assert_equals: expected 216 but got 223
FAIL spaces linebreak spaces assert_equals: expected 216 but got 223
FAIL multiple linebreaks assert_equals: expected 216 but got 223
FAIL multiple linebreaks + spaces assert_equals: expected 216 but got 223
Harness: the test ran to completion.
